Adding option to monitor specific elements (#73)
A list of element tags is supplied by the user to plot_deformedshape() function. This will plot the deformed shape of the specified elements only. Useful for monitoring deformation of specific locations in large structures since Matplotlib 3D plot can not zoom in to a specific location. It was helpful for my research on rocking bridge columns.